+++ crates/burn-jit/src/kernel/conv/conv2d/mod.rs
@@ -5,7 +5,6 @@ mod im2col;
 mod implicit_gemm;
 mod transpose_direct;
 
-#[cfg(feature = "autotune")]
 mod tune;
 
 pub use base::*;
@@ -14,5 +13,4 @@ pub use direct::*;
 pub use im2col::*;
 pub use implicit_gemm::*;
 pub use transpose_direct::*;
-#[cfg(feature = "autotune")]
 pub use tune::*;
