+++ crates/burn-jit/src/kernel/conv/conv2d/tune/conv2d.rs
@@ -5,9 +5,7 @@ use burn_tensor::{
 use cubecl::{
     tune,
     tune::{local_tuner, tune_with, LocalTuner},
-    AutotuneKey,
 };
-use serde::{Deserialize, Serialize};
 
 use crate::{
     kernel::{
@@ -17,6 +15,8 @@ use crate::{
     tensor::JitTensor,
     FloatElement, IntElement, JitAutotuneKey, JitRuntime, JitTuneId,
 };
+
+use super::Conv2dAutotuneKey;
 
 /// Executes autotune on conv2d operations
 pub fn conv2d_autotune<R: JitRuntime, E: FloatElement, I: IntElement>(
@@ -36,27 +36,6 @@ pub fn conv2d_autotune<R: JitRuntime, E: FloatElement, I: IntElement>(
             input, weights, bias, options,
         )),
     )
-}
-
-#[derive(Hash, Eq, PartialEq, Debug, Clone, Serialize, Deserialize, AutotuneKey)]
-/// Autotune key representative of matmul versions
-pub struct Conv2dAutotuneKey {
-    pub kernel_size: [usize; 2],
-    pub stride: [usize; 2],
-    pub padding: [usize; 2],
-    pub dilation: [usize; 2],
-    pub groups: usize,
-    #[autotune(anchor)]
-    pub in_channels: usize,
-    #[autotune(anchor)]
-    pub out_channels: usize,
-    #[autotune(anchor)]
-    pub height: usize,
-    #[autotune(anchor)]
-    pub width: usize,
-    #[autotune(anchor)]
-    pub batch_size: usize,
-    pub has_bias: bool,
 }
 
 #[tune(
