+++ crates/burn-jit/src/kernel/conv/conv2d/tune/mod.rs
@@ -1,5 +1,12 @@
+#[cfg(feature = "autotune")]
 mod conv2d;
+#[cfg(feature = "autotune")]
 mod conv_transpose2d;
 
+#[cfg(feature = "autotune")]
 pub use conv2d::*;
+#[cfg(feature = "autotune")]
 pub use conv_transpose2d::*;
+
+mod key;
+pub use key::*;
